% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blant.R
\name{blant}
\alias{blant}
\title{The BLANT function}
\usage{
blant(filepath, mode, n, k, filename, connected, threads = -1)
}
\arguments{
\item{mode}{--output Mode: o (ODV, the default); i (indexGraphlets); g (GDV); f (graphletFrequency)}

\item{n}{-- Sampling number}

\item{k}{--The number of nodes in graphlets to be sampled}

\item{filename}{--Graph must be in one of the following formats with its extension name:
GML (.gml) GraphML (.xml) LGF(.lgf) CSV(.csv) LEDA(.leda) Edgelist (.el) .}

\item{connected}{-- if it is TRUE/T, blant function will return a table ONLY contains connected graphlets.
Otherwise, it can contain both connected and disconnected graphlets.}

\item{filepath:}{The path to your BLANT function}
}
\description{
This function will return a table containing the sampling results\cr\cr\cr
NOTE: \cr\cr
Before using this package, MAKE SURE you have a working Blant function installed\cr\cr
1. Nodes must be integers numbered 0 through n-1, inclusive.\cr\cr
2. Duplicates and self-loops should be removed before calling BLANT.
}
\examples{
blant(filepath=..., mode="f", n=10000, k=5, filename="syeast.el", connected=T)
}
\keyword{blant}
