% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blant.R
\name{blant}
\alias{blant}
\title{The blant function}
\usage{
blant(mode, n, k, filename, threads = -1)
}
\arguments{
\item{mode}{--output Mode: o (ODV, the default); i (indexGraphlets); g (GDV); f (graphletFrequency)}

\item{n}{-- Sampling number}

\item{k}{--The number of nodes in graphlets to be sampled}

\item{filename}{--Graph must be in one of the following formats with its extension name:
GML (.gml) GraphML (.xml) LGF(.lgf) CSV(.csv) LEDA(.leda) Edgelist (.el) .}
}
\description{
This function will return a table containing the sampling results\cr
Note:  1. Nodes must be integers numbered 0 through n-1, inclusive.\cr
2. Duplicates and self-loops should be removed before calling BLANT.
}
\keyword{blant}
